Adding additional properties and fix turnover type

+++ dotnet/Qred.Connect/Models/ApplicationDecisionApproved.cs
@@ -5,7 +5,7 @@ namespace Qred.Connect
     public class ApplicationDecisionApproved : ApplicationDecision
     {
         public decimal? Amount { get; set; }
-        public decimal? Term { get; set; }
+        public int? Term { get; set; }
         public string Campaign { get; set; }
         public Uri OfferUrl { get; set; }
